fix(theme): opaque Applications dropdown popovers + rendered audit evidence (#14232)
The #14236 retokenization replaced the hard-dark bg-neutral-800 Applications
select/tooltip popovers with bg-popover, but --popover is undefined across the
entire UI token set (no --popover: rule in any theme in the built CSS), so
bg-popover resolves to a transparent rgba(0,0,0,0) surface. On the always-dark
theme-cloud console that is a regression: menu items float over whatever is
behind them.

Retokenize the four flagged Applications popovers to bg-card (defined in every
theme; = brand-black under theme-cloud) — matching the SelectTrigger and the
majority of sibling working cloud dropdowns. Scope kept to the Applications
surfaces #14232 named; the same dead-token in organization/api-keys is noted as
a follow-up.

Evidence:
- Unblocked the app/cloud aesthetic audit (the prior audit:app failure was a
  missing bun install + unbuilt workspace /browser dists, not a code problem);
  audit:cloud now produces aesthetic-audit-output-cloud (68 passed; the 1
  failure is pre-existing route-coverage drift unrelated to this change).
- New audit-app-dropdown Playwright project opens the touched range Selects in
  both app theme modes and asserts opacity + bg/text contrast (4 passed).
- Rendered proof committed under .github/issue-evidence/14232-*.

Verification: touched-component vitest 6/6, settings token guard 2/2,
build:client 104/104.

Fixes #14232
